Credentials are optional
With sbt publishing, credentials are only required if the endpoint receiving the artifacts requires them. This commit therefore retains this behaviour. As a consequence, the plugin may be used in situations where the user has no credentials given that they are not authorised to publish artifacts. A valid scenario is that Continuous Delivery is authorised to publish, but not an individual developer.